laid the ground work for supporting both the native API and backwards compatibility with old versions
